refactor: move Log config to Hoard.Effects.Log
Helps prevent a cyclic dependency in the Peer Manager PR I'm working on,
and it makes sense to let `Log` own its own config.
