+++ src/Hoard/Types/Environment.hs
@@ -1,7 +1,5 @@
 module Hoard.Types.Environment
     ( ServerConfig (..)
-    , LogConfig (..)
-    , Severity (..)
     , Config (..)
     , Handles (..)
     , Topology (..)
@@ -30,7 +28,6 @@ module Hoard.Types.Environment
     , NodeSocketsConfig (..)
     , SshTunnel (..)
     , Local (..)
-    , defaultLogConfig
     )
 where
 
@@ -42,11 +39,11 @@ import Ouroboros.Network.IOManager (IOManager)
 import Hoard.BlockFetch.Config qualified as BlockFetch
 import Hoard.ChainSync.Config qualified as ChainSync
 import Hoard.Collectors.Config qualified as Collectors
+import Hoard.Effects.Log qualified as Log
 import Hoard.KeepAlive.Config qualified as KeepAlive
 import Hoard.PeerSharing.Config qualified as PeerSharing
 import Hoard.Types.Cardano (CardanoBlock)
 import Hoard.Types.DBConfig (DBPools (..))
-import Hoard.Types.JsonReadShow (JsonReadShow (..))
 import Hoard.Types.QuietSnake (QuietSnake (..))
 
 
@@ -59,34 +56,11 @@ data ServerConfig = ServerConfig
     deriving (FromJSON) via QuietSnake ServerConfig
 
 
-data LogConfig = LogConfig
-    { minimumSeverity :: Severity
-    , output :: Handle
-    }
-
-
-data Severity
-    = DEBUG
-    | INFO
-    | WARN
-    | ERROR
-    deriving stock (Eq, Ord, Enum, Bounded, Show, Read)
-    deriving (FromJSON) via (JsonReadShow Severity)
-
-
-defaultLogConfig :: LogConfig
-defaultLogConfig =
-    LogConfig
-        { minimumSeverity = minBound
-        , output = stdout
-        }
-
-
 -- | Pure configuration data loaded from config files
 data Config = Config
     { server :: ServerConfig
     , nodeSockets :: NodeSocketsConfig
-    , logging :: LogConfig
+    , logging :: Log.Config
     , protocolInfo :: ProtocolInfo CardanoBlock
     , nodeConfig :: NodeConfig
     , maxFileDescriptors :: Maybe Word32
